Refactor project structure to MVVM
Organizing the project using MVVM will separate the user interface from the
business logic and data handling.

By doing so, It will be easier for us to see what's going around the project
instead of putting all stuff into `main.dart`.
