+++ lib/viewmodels/theme_viewmodel.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class ThemeViewModel extends ChangeNotifier {
+  var _themeMode = ThemeMode.system;
+
+  ThemeMode get themeMode => _themeMode;
+  set themeMode(ThemeMode value) {
+    _themeMode = value;
+
+    notifyListeners();
+  }
+
+  get darkTheme => ThemeData(
+    useMaterial3: true,
+    colorScheme: ColorScheme.fromSeed(
+      seedColor: Colors.amber,
+      brightness: Brightness.dark,
+    ),
+  );
+
+  get lightTheme => ThemeData(
+    useMaterial3: true,
+    colorScheme: ColorScheme.fromSeed(
+      seedColor: Colors.amber,
+      brightness: Brightness.light,
+    ),
+  );
+}
